Shuffle around a few functions.
