+++ jaq-parse/src/lib.rs
@@ -10,6 +10,7 @@ mod filter;
 mod path;
 mod prec_climb;
 mod token;
+mod string;
 
 use jaq_syn as syn;
 
@@ -22,17 +23,6 @@ use syn::Spanned;
 
 /// Lex/parse error.
 pub type Error = Simple<String>;
-
-/// A (potentially empty) parenthesised and `;`-separated sequence of arguments.
-fn args<T, P>(arg: P) -> impl Parser<Token, Vec<T>, Error = P::Error> + Clone
-where
-    P: Parser<Token, T> + Clone,
-{
-    arg.separated_by(just(Token::Ctrl(';')))
-        .delimited_by(just(Token::Ctrl('(')), just(Token::Ctrl(')')))
-        .or_not()
-        .map(Option::unwrap_or_default)
-}
 
 fn lex() -> impl Parser<char, Vec<Spanned<Token>>, Error = Simple<char>> {
     recursive(token::tree)
